pyproject.toml
[tool.poetry]
name = "email-bot"
version = "0.1.0"
description = "IU3"
package-mode = false
authors = ["snail <sgefsnkiospch908@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.13"
aiogram = "^3.22.0"
python-decouple = "^3.8"
redis = "^6.4.0"
django = "^5.2.7"
psycopg2-binary = "^2.9.10"
djangorestframework = "^3.16.1"

[tool.poetry.group.dev.dependencies]
black = "^25.9.0"
isort = "^6.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"